Update the README so that it's cut & paste-able
In the README the example shows how to tweet (or status update, as
it's officially known), which doesn't work as is.

Because the credentials are passed in as '*creds*' which is not declared
anywhere.

But 'my-creds' is.

Applying this change so that anybody reading the README can just cut &
paste the example and have it working.

There is nothing like instant gratification to 'sell' the library.