Updated the instructions for getting the API keys
In the previous write-up, the URL used was:

https://dev.twitter.com/apps/<app-id>

which would show up as

https://dev.twitter.com/apps/ once the Markdown was rendered in the
browser (note the omitted '<app-id>'!).

The same goes for this URL:

https://dev.twitter.com/apps/<app-id>/my_token

it was rendered as:

https://dev.twitter.com/apps//my_token

(note the double slash '//'!).

If you clicked on those links, you'd get an error.

My instructions work around the above issue by providing step by step
instructions for the potential library users.